+++ pkg/builder/controller.go
@@ -18,12 +18,9 @@ package builder
 
 import (
 	"fmt"
-	"net/http"
-	"net/url"
 	"strings"
 
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
@@ -33,8 +30,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 	"sigs.k8s.io/controller-runtime/pkg/source"
-	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
-	"sigs.k8s.io/controller-runtime/pkg/webhook/conversion"
 )
 
 // Supporting mocking out functions for testing
@@ -71,8 +66,6 @@ func ControllerManagedBy(m manager.Manager) *Builder {
 // update events by *reconciling the object*.
 // This is the equivalent of calling
 // Watches(&source.Kind{Type: apiType}, &handler.EnqueueRequestForObject{})
-// If the passed in object has implemented the admission.Defaulter interface, a MutatingWebhook will be wired for this type.
-// If the passed in object has implemented the admission.Validator interface, a ValidatingWebhook will be wired for this type.
 //
 // Deprecated: Use For
 func (blder *Builder) ForType(apiType runtime.Object) *Builder {
@@ -83,8 +76,6 @@ func (blder *Builder) ForType(apiType runtime.Object) *Builder {
 // update events by *reconciling the object*.
 // This is the equivalent of calling
 // Watches(&source.Kind{Type: apiType}, &handler.EnqueueRequestForObject{})
-// If the passed in object has implemented the admission.Defaulter interface, a MutatingWebhook will be wired for this type.
-// If the passed in object has implemented the admission.Validator interface, a ValidatingWebhook will be wired for this type.
 func (blder *Builder) For(apiType runtime.Object) *Builder {
 	blder.apiType = apiType
 	return blder
@@ -159,7 +150,7 @@ func (blder *Builder) Build(r reconcile.Reconciler) (manager.Manager, error) {
 	}
 
 	// Set the Config
-	if err := blder.doConfig(); err != nil {
+	if err := blder.loadRestConfig(); err != nil {
 		return nil, err
 	}
 
@@ -170,11 +161,6 @@ func (blder *Builder) Build(r reconcile.Reconciler) (manager.Manager, error) {
 
 	// Set the ControllerManagedBy
 	if err := blder.doController(r); err != nil {
-		return nil, err
-	}
-
-	// Set the Webook if needed
-	if err := blder.doWebhook(); err != nil {
 		return nil, err
 	}
 
@@ -217,7 +203,7 @@ func (blder *Builder) doWatch() error {
 	return nil
 }
 
-func (blder *Builder) doConfig() error {
+func (blder *Builder) loadRestConfig() error {
 	if blder.config != nil {
 		return nil
 	}
@@ -257,80 +243,4 @@ func (blder *Builder) doController(r reconcile.Reconciler) error {
 	}
 	blder.ctrl, err = newController(name, blder.mgr, controller.Options{Reconciler: r})
 	return err
-}
-
-func (blder *Builder) doWebhook() error {
-	// Create a webhook for each type
-	gvk, err := apiutil.GVKForObject(blder.apiType, blder.mgr.GetScheme())
-	if err != nil {
-		return err
-	}
-
-	// TODO: When the conversion webhook lands, we need to handle all registered versions of a given group-kind.
-	// A potential workflow for defaulting webhook
-	// 1) a bespoke (non-hub) version comes in
-	// 2) convert it to the hub version
-	// 3) do defaulting
-	// 4) convert it back to the same bespoke version
-	// 5) calculate the JSON patch
-	//
-	// A potential workflow for validating webhook
-	// 1) a bespoke (non-hub) version comes in
-	// 2) convert it to the hub version
-	// 3) do validation
-	if defaulter, isDefaulter := blder.apiType.(admission.Defaulter); isDefaulter {
-		mwh := admission.DefaultingWebhookFor(defaulter)
-		if mwh != nil {
-			path := generateMutatePath(gvk)
-
-			// Checking if the path is already registered.
-			// If so, just skip it.
-			if !blder.isAlreadyHandled(path) {
-				log.Info("Registering a mutating webhook",
-					"GVK", gvk,
-					"path", path)
-				blder.mgr.GetWebhookServer().Register(path, mwh)
-			}
-		}
-	}
-
-	if validator, isValidator := blder.apiType.(admission.Validator); isValidator {
-		vwh := admission.ValidatingWebhookFor(validator)
-		if vwh != nil {
-			path := generateValidatePath(gvk)
-
-			// Checking if the path is already registered.
-			// If so, just skip it.
-			if !blder.isAlreadyHandled(path) {
-				log.Info("Registering a validating webhook",
-					"GVK", gvk,
-					"path", path)
-				blder.mgr.GetWebhookServer().Register(path, vwh)
-			}
-		}
-	}
-
-	err = conversion.CheckConvertibility(blder.mgr.GetScheme(), blder.apiType)
-	if err != nil {
-		log.Error(err, "conversion check failed", "GVK", gvk)
-	}
-	return nil
-}
-
-func (blder *Builder) isAlreadyHandled(path string) bool {
-	h, p := blder.mgr.GetWebhookServer().WebhookMux.Handler(&http.Request{URL: &url.URL{Path: path}})
-	if p == path && h != nil {
-		return true
-	}
-	return false
-}
-
-func generateMutatePath(gvk schema.GroupVersionKind) string {
-	return "/mutate-" + strings.Replace(gvk.Group, ".", "-", -1) + "-" +
-		gvk.Version + "-" + strings.ToLower(gvk.Kind)
-}
-
-func generateValidatePath(gvk schema.GroupVersionKind) string {
-	return "/validate-" + strings.Replace(gvk.Group, ".", "-", -1) + "-" +
-		gvk.Version + "-" + strings.ToLower(gvk.Kind)
 }
